requirements.txt
asgiref==3.2.9
cached-property==1.5.1
Django==3.0.7
django-cors-headers==3.4.0
django-url-filter==0.3.15
djangorestframework==3.11.0
enum-compat==0.0.3
mysqlclient==1.4.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
